+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Settings will be saved periodically when using the app
 - Add `save` button to settings window
 - Add ability to change docker socket location in settings
+- Display a confirmation popup when deleting a container/image